fixed current user stuff
