+++ app/src/main/java/edu/ucsb/cs/cs184/georgelieu/thingamajob/LoginActivity.java
@@ -51,6 +51,7 @@ import java.util.Arrays;
 
 public class LoginActivity extends AppCompatActivity {
 
+    public static User current_user = new User("", "", "");
     private EditText inputEmail, inputPassword;
     private FirebaseAuth auth;
     private ProgressBar progressBar;
@@ -107,7 +108,7 @@ public class LoginActivity extends AppCompatActivity {
                                             @Override
                                             public void onDataChange(DataSnapshot dataSnapshot) {
                                                 for (DataSnapshot child : dataSnapshot.getChildren()) {
-                                                    queryUserInfo(child.getKey());
+                                                    MainActivity.queryUserInfo(child.getKey());
                                                 }
                                             }
                                             @Override
@@ -206,43 +207,10 @@ public class LoginActivity extends AppCompatActivity {
                                 }
                             }
                         });
-                // query database for user info to set current_user object in DatabaseHelper
-                Query current_user = DatabaseHelper.mDatabaseReference.child("Users").orderByChild("email").equalTo(email);
-
-                current_user.addValueEventListener( new ValueEventListener(){
-                    @Override
-                    public void onDataChange(DataSnapshot dataSnapshot) {
-                        for (DataSnapshot child : dataSnapshot.getChildren()) {
-                            queryUserInfo(child.getKey());
-                        }
-                    }
-
-                    @Override
-                    public void onCancelled(DatabaseError databaseError) {}
-                });
             }
         });
     }
 
-    public static void queryUserInfo(final String userId){
-        Query user_info = DatabaseHelper.mDatabaseReference.child("Users").child(userId);
-        user_info.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                String email = dataSnapshot.child("email").getValue().toString();
-                String fullname = dataSnapshot.child("full_name").getValue().toString();
-
-                DatabaseHelper.current_user.setUser_id(userId);
-                DatabaseHelper.current_user.setEmail(email);
-                DatabaseHelper.current_user.setFull_name(fullname);
-
-            }
-
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-            }
-        });
-    }
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
