+++ app/src/main/res/layout/activity_maps.xml
@@ -24,5 +24,14 @@
         android:layout_alignParentRight="true"
         android:layout_margin="8dp"/>
 
+    <Button
+        android:id="@+id/sign_out_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_alignParentTop="true"
+        android:layout_marginStart="12dp"
+        android:text="sign_out" />
+
 
 </RelativeLayout>
